Substantially changed plan evaluator
1. Introduced a seperate representations for the join
orders. Each join order plan is now a binary tree
and we store the binary tree along with the evaluated
cost in the DP Table.
2. The cost evaluation mentioned in step 1 is done in
a referentially transparent manner. All side-effects to the
root within the parallel_eval and restored. I'm hoping this
is the only place where the side effects occur. If not,
my claim about referential transparency is wrong.